app/requirements.txt
streamlit>=1.32.0
opencv-python-headless
openai>=1.0.0
chromadb==0.4.24
networkx
plotly
numpy<2.0
Pillow
python-dotenv
scipy
openai-whisper
moviepy
ffmpeg-python
torch
torchvision
scikit-learn
scikit-image